+++ core/entity/task/template.js
@@ -1,7 +1,12 @@
 var mongodb = require('../../lib/mongodb').db;
 var BaseSchema = require('./schema');
+var Script = require('../script').Entity;
 
-var TemplateSchema = BaseSchema.EntitySchema.extend({ });
+var TemplateSchema = BaseSchema.EntitySchema.extend({
+  script_id : { type: String, ref: 'Script' },
+  script_arguments : { type: Array, 'default': [] },
+  script_runas : { type: String, 'default':'' },
+});
 
 TemplateSchema.methods.values = function(){
   var template = this
@@ -11,6 +16,16 @@ TemplateSchema.methods.values = function(){
 		'script_id': template.script_id,
 		'script_arguments': template.script_arguments,
   }
+}
+
+TemplateSchema.methods.publish = function(done){
+  var data = this.toObject();
+  if( ! this.script_id ) return done();
+  Script.findById(this.script_id, function(err,script){
+    if(err||!script) return done(data);
+    data.script_name = script.filename;
+    done(data);
+  });
 }
 
 var updateFn = TemplateSchema.methods.update;